refactor : 순환 참조 해결
