+++ src/main/java/dbdr/global/util/ApiUtils.java
@@ -1,4 +1,4 @@
-package dbdr.util;
+package dbdr.global.util;
 
 import org.springframework.http.HttpStatus;
